backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
httpx==0.27.0
huggingface-hub>=0.34.0,<1.0
groq==0.11.0
python-multipart==0.0.9
pydantic>=2.0.0
torch>=2.0.0
transformers>=4.40.0
pillow>=10.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
